requirements.txt
# Video Inspection System - Requirements
# ======================================
# Architecture: YOLO Detection → Frame Gating → Claude Vision → Rule Engine

# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0

# Object detection
ultralytics>=8.0.0  # YOLOv8

# Claude API for Vision analysis
anthropic>=0.18.0

# System monitoring (optional, for test.py)
psutil>=5.9.0

# Image processing
Pillow>=10.0.0

# HTTP client for local model services
requests>=2.31.0
